Add UIViewRepresentable for SwiftUI.
